refact(decman-lib): move support types and validators
The payload support types move to the catalog with feature-gated
utoipa/ts-rs derives; decman re-exports them so paths and the HTTP
JSON stay identical. The protocol validators move to the framework;
validate_future_micros now takes now_micros as a parameter, and every
message string stays byte-identical inside Error::Validation.
