+++ crates/decman-lib/src/catalog/mod.rs
@@ -1,1 +1,3 @@
 //! DecMan's protocol content.
+
+pub mod types;
