+++ crates/decman/src/server/types.rs
@@ -46,6 +46,13 @@ pub use common::types::{
     PartyMetadata, PeerErrorKind, PeerPackageComparison, PeerPackageResult, PendingInvitation,
     Permission, VettedPackageInfo, WorkflowKind, WorkflowProgress, WorkflowRole, WorkflowRun,
 };
+// The payload support types and their protocol validators now live in
+// `decman-lib`, re-exported here so existing `crate::server::types::X`
+// paths keep resolving unchanged.
+pub use decman_lib::catalog::types::{
+    AppRewardBeneficiary, BillingParams, FarConfig, RewardBeneficiary, VaultLimits,
+};
+use decman_lib::framework::validate::*;
 
 use crate::{canton_id::CantonId, noise::server::ActiveWorkflow};
 
@@ -315,51 +322,9 @@ pub type OnboardingResponse = WorkflowResponse;
 // Governance Types (Structured Actions)
 // ============================================================================
 
-/// Vault limits configuration (all fields are optional in Daml)
-#[derive(Clone, Debug, Serialize, Deserialize, utoipa::ToSchema)]
-#[cfg_attr(feature = "typegen", derive(ts_rs::TS), ts(optional_fields))]
-pub struct VaultLimits {
-    #[schema(value_type = Option<String>)]
-    #[cfg_attr(feature = "typegen", ts(type = "string"))]
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub max_total_deposit: Option<DamlDecimal>,
-    #[schema(value_type = Option<String>)]
-    #[cfg_attr(feature = "typegen", ts(type = "string"))]
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub min_deposit_amount: Option<DamlDecimal>,
-    #[schema(value_type = Option<String>)]
-    #[cfg_attr(feature = "typegen", ts(type = "string"))]
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub min_withdrawal_amount: Option<DamlDecimal>,
-}
-
-/// Featured App Right beneficiary
-#[derive(Clone, Debug, Serialize, Deserialize, utoipa::ToSchema)]
-#[cfg_attr(feature = "typegen", derive(ts_rs::TS), ts(optional_fields))]
-pub struct AppRewardBeneficiary {
-    pub beneficiary: CantonId,
-    #[schema(value_type = String)]
-    #[cfg_attr(feature = "typegen", ts(type = "string"))]
-    pub weight: DamlDecimal,
-}
-
-/// A CIP-104 reward-coupon beneficiary assignment.
-#[derive(Clone, Debug, Serialize, Deserialize, utoipa::ToSchema)]
-#[cfg_attr(feature = "typegen", derive(ts_rs::TS), ts(optional_fields))]
-pub struct RewardBeneficiary {
-    pub beneficiary: CantonId,
-    #[schema(value_type = String)]
-    #[cfg_attr(feature = "typegen", ts(type = "string"))]
-    pub percentage: DamlDecimal,
-}
-
-/// Featured App Right configuration
-#[derive(Clone, Debug, Serialize, Deserialize, utoipa::ToSchema)]
-#[cfg_attr(feature = "typegen", derive(ts_rs::TS), ts(optional_fields))]
-pub struct FarConfig {
-    pub featured_app_right_cid: String,
-    pub beneficiaries: Vec<AppRewardBeneficiary>,
-}
+// `VaultLimits`, `AppRewardBeneficiary`, `RewardBeneficiary`, and
+// `FarConfig` now live in `decman_lib::catalog::types` and are re-exported
+// above so existing `crate::server::types::X` paths keep resolving.
 
 /// Structured action types for Vault governance
 #[derive(Clone, Debug, Serialize, Deserialize, utoipa::ToSchema)]
@@ -489,11 +454,11 @@ impl ActionType {
             ActionType::GovernanceAddMember { new_threshold, .. }
             | ActionType::GovernanceRemoveMember { new_threshold, .. }
             | ActionType::GovernanceSetThreshold { new_threshold } => {
-                validate_threshold(*new_threshold)
+                validate_threshold(*new_threshold).map_err(|e| e.to_string())
             }
             ActionType::GovernanceSetTimeout {
                 new_timeout_microseconds,
-            } => validate_timeout(*new_timeout_microseconds),
+            } => validate_timeout(*new_timeout_microseconds).map_err(|e| e.to_string()),
             ActionType::VaultDeployment {
                 vault_far_config: Some(far),
                 ..
@@ -501,130 +466,21 @@ impl ActionType {
             | ActionType::ProcessorDeploymentRequest {
                 processor_far_config: Some(far),
                 ..
-            } => validate_beneficiary_weights(&far.beneficiaries),
+            } => validate_beneficiary_weights(&far.beneficiaries).map_err(|e| e.to_string()),
             ActionType::VaultUpdateFarBeneficiaries {
                 new_beneficiaries, ..
-            } => validate_beneficiary_weights(new_beneficiaries),
+            } => validate_beneficiary_weights(new_beneficiaries).map_err(|e| e.to_string()),
             _ => Ok(()),
         }
     }
 }
 
-fn validate_threshold(new_threshold: i64) -> Result<(), String> {
-    if new_threshold < 1 {
-        return Err(format!(
-            "new_threshold must be at least 1, got {new_threshold}"
-        ));
-    }
-    Ok(())
-}
-
-fn validate_timeout(microseconds: i64) -> Result<(), String> {
-    if microseconds <= 0 {
-        return Err(format!(
-            "new_timeout_microseconds must be positive, got {microseconds}"
-        ));
-    }
-    Ok(())
-}
-
-fn validate_unique_issuers(issuers: &[CantonId], field: &str) -> Result<(), String> {
-    let mut seen = HashSet::new();
-    for issuer in issuers {
-        if !seen.insert(issuer) {
-            return Err(format!("{field} must not list {issuer} more than once"));
-        }
-    }
-    Ok(())
-}
-
-/// Mirrors the Daml `selfIssuedRequirementsHaveClaims` guard. A requirement the
-/// governance party issues itself must name at least one claim. The mint
-/// refuses a claimless self-issued credential, because it attests for nobody.
-/// Requirements from other issuers are out of scope: those credentials arrive
-/// out of band.
-fn validate_self_issued_requirements_have_claims(
-    requirements: &[PartyCredentialRequirement],
-    governance_party: &CantonId,
-    field: &str,
-) -> Result<(), String> {
-    for requirement in requirements {
-        if requirement.issuer == *governance_party && requirement.required_claims.is_empty() {
-            return Err(format!(
-                "{field}: a requirement issued by the governance party must list at least one required claim"
-            ));
-        }
-    }
-    Ok(())
-}
-
-/// Reject an epoch-microsecond instant that is not in the future.
-///
-/// The on-ledger `executeImpl` asserts the same thing, but only at execute
-/// time — after a full propose/confirm round has been spent on a value that
-/// could never have worked.
-fn validate_future_micros(micros: i64, field: &str) -> Result<(), String> {
-    if micros <= 0 {
-        return Err(format!("{field} must be positive, got {micros}"));
-    }
-    let now_micros = Utc::now().timestamp_micros();
-    if micros <= now_micros {
-        return Err(format!(
-            "{field} must be in the future, got {micros} (now {now_micros})"
-        ));
-    }
-    Ok(())
-}
-
-fn validate_positive_amount(amount: &DamlDecimal, field: &str) -> Result<(), String> {
-    // `DamlDecimal` itself doesn't implement `PartialOrd`; compare via the
-    // inner `rust_decimal::Decimal` returned by `value()` against a parsed
-    // zero so we don't need a direct dep on `rust_decimal`.
-    let zero = "0"
-        .parse::<DamlDecimal>()
-        .expect("'0' is a valid DamlDecimal")
-        .value();
-    if amount.value() <= zero {
-        return Err(format!("{field} must be strictly positive, got {amount}"));
-    }
-    Ok(())
-}
-
-fn validate_beneficiary_weights(beneficiaries: &[AppRewardBeneficiary]) -> Result<(), String> {
-    if beneficiaries.is_empty() {
-        return Ok(());
-    }
-    let sum: DamlDecimal = beneficiaries.iter().map(|b| b.weight).sum();
-    let one: DamlDecimal = "1".parse().expect("'1' is a valid DamlDecimal");
-    if sum != one {
-        return Err(format!(
-            "FAR beneficiary weights must sum to exactly 1.0, got {sum}"
-        ));
-    }
-    Ok(())
-}
-
-/// Billing parameters for a paid credential.
-/// Mirrors `Utility.Credential.App.V0.Types.BillingParams`.
-#[derive(Clone, Debug, Serialize, Deserialize, utoipa::ToSchema)]
-#[cfg_attr(feature = "typegen", derive(ts_rs::TS), ts(optional_fields))]
-pub struct BillingParams {
-    /// The daily fee for the credential in USD (corresponds to RatePerDay record).
-    #[schema(value_type = String)]
-    #[cfg_attr(feature = "typegen", ts(type = "string"))]
-    pub fee_per_day_usd: DamlDecimal,
-    /// Duration between fee charges, in minutes.
-    pub billing_period_minutes: i64,
-    /// Target deposit amount in USD.
-    #[schema(value_type = String)]
-    #[cfg_attr(feature = "typegen", ts(type = "string"))]
-    pub deposit_target_amount_usd: DamlDecimal,
-    /// Holder's weight on the activity marker (0.0 - 1.0). None means 0.
-    #[serde(default)]
-    #[schema(value_type = Option<String>)]
-    #[cfg_attr(feature = "typegen", ts(type = "string"))]
-    pub holder_activity_weight: Option<DamlDecimal>,
-}
+// The protocol validators (`validate_threshold`, `validate_timeout`,
+// `validate_unique_issuers`, `validate_self_issued_requirements_have_claims`,
+// `validate_future_micros`, `validate_positive_amount`,
+// `validate_beneficiary_weights`, `validate_reward_beneficiaries`) and
+// `BillingParams` now live in `decman-lib` (`framework::validate` /
+// `catalog::types`), imported and re-exported above.
 
 /// Types of governance domain action proposals
 #[derive(Clone, Debug, Deserialize, Serialize, utoipa::ToSchema)]
@@ -908,19 +764,21 @@ impl ProposalType {
                 validity_window_hours,
                 ..
             } => {
-                validate_positive_amount(amount, "amount")?;
+                validate_positive_amount(amount, "amount").map_err(|e| e.to_string())?;
                 if *validity_window_hours == Some(0) {
                     return Err("validity_window_hours must be greater than 0".to_string());
                 }
                 Ok(())
             }
             ProposalType::Mint { amount, .. } | ProposalType::Burn { amount, .. } => {
-                validate_positive_amount(amount, "amount")
+                validate_positive_amount(amount, "amount").map_err(|e| e.to_string())
             }
             ProposalType::OfferPaidCredential {
                 deposit_initial_amount_usd: Some(d),
                 ..
-            } => validate_positive_amount(d, "deposit_initial_amount_usd"),
+            } => {
+                validate_positive_amount(d, "deposit_initial_amount_usd").map_err(|e| e.to_string())
+            }
             ProposalType::SetupMintingDelegation {
                 expires_at_micros,
                 amulet_merge_limit,
@@ -929,7 +787,12 @@ impl ProposalType {
                 if *amulet_merge_limit <= 0 {
                     return Err("amulet_merge_limit must be greater than 0".to_string());
                 }
-                validate_future_micros(*expires_at_micros, "expires_at_micros")
+                validate_future_micros(
+                    *expires_at_micros,
+                    Utc::now().timestamp_micros(),
+                    "expires_at_micros",
+                )
+                .map_err(|e| e.to_string())
             }
             ProposalType::AcceptExternalPartySetup { proposal_cid } => {
                 if proposal_cid.trim().is_empty() {
@@ -940,7 +803,7 @@ impl ProposalType {
             ProposalType::SetProviderAppRewardBeneficiaries {
                 provider_app_reward_beneficiaries: Some(beneficiaries),
                 ..
-            } => validate_beneficiary_weights(beneficiaries),
+            } => validate_beneficiary_weights(beneficiaries).map_err(|e| e.to_string()),
             // Mirrors the template's `ensure` guard: onboarding zero issuers
             // does no work, and a duplicated issuer would mint two
             // credentials sharing one id. Reject both with a 400 before the
@@ -952,6 +815,7 @@ impl ProposalType {
                     return Err("instrument_issuers must not be empty".to_string());
                 }
                 validate_unique_issuers(instrument_issuers, "instrument_issuers")
+                    .map_err(|e| e.to_string())
             }
             ProposalType::ProvisionInstrument {
                 initial_instrument_issuers,
@@ -962,8 +826,10 @@ impl ProposalType {
                     issuer_requirements,
                     governance_party,
                     "issuer_requirements",
-                )?;
+                )
+                .map_err(|e| e.to_string())?;
                 validate_unique_issuers(initial_instrument_issuers, "initial_instrument_issuers")
+                    .map_err(|e| e.to_string())
             }
             ProposalType::CreateProviderConfiguration {
                 registrar_requirements,
@@ -972,7 +838,8 @@ impl ProposalType {
                 registrar_requirements,
                 governance_party,
                 "registrar_requirements",
-            ),
+            )
+            .map_err(|e| e.to_string()),
             ProposalType::OffboardInstrumentIssuers { instrument_issuers } => {
                 if instrument_issuers.is_empty() {
                     return Err("instrument_issuers must not be empty".to_string());
@@ -1014,7 +881,7 @@ impl ProposalType {
                         return Err(format!("duplicate assigner not allowed: {a}"));
                     }
                 }
-                validate_reward_beneficiaries(new_beneficiaries)
+                validate_reward_beneficiaries(new_beneficiaries).map_err(|e| e.to_string())
             }
             ProposalType::RevokeCouponReassignmentDelegation { delegation } => {
                 if delegation.trim().is_empty() {
@@ -1025,58 +892,6 @@ impl ProposalType {
             _ => Ok(()),
         }
     }
-}
-
-/// Validates a `new_beneficiaries` list (e.g.
-/// `SetupCouponReassignmentDelegation::new_beneficiaries`): non-empty,
-/// <= 20 entries, no duplicate beneficiary, each percentage in (0.0, 1.0],
-/// summing to exactly 1.0.
-///
-/// The uniqueness rule mirrors the on-ledger `RewardCoupon_AssignBeneficiaries`
-/// impl (`require "Beneficaries are unique"`); catching it here means a
-/// duplicated split is rejected at propose time rather than passing the vote
-/// and then failing every `Delegation_Assign`, which would leave a permanently
-/// unusable delegation.
-///
-/// `DamlDecimal` addition is exact (no float rounding), so an exact `==`
-/// against `1.0` is sufficient here — no epsilon tolerance is needed.
-fn validate_reward_beneficiaries(beneficiaries: &[RewardBeneficiary]) -> Result<(), String> {
-    if beneficiaries.is_empty() {
-        return Err("new_beneficiaries must not be empty".to_string());
-    }
-    if beneficiaries.len() > 20 {
-        return Err("at most 20 beneficiaries per coupon".to_string());
-    }
-    let one = DamlDecimal::parse("1").map_err(|e| e.to_string())?;
-    let mut seen = std::collections::HashSet::new();
-    for b in beneficiaries {
-        if b.percentage.value() <= DamlDecimal::ZERO.value() || b.percentage.value() > one.value() {
-            return Err(format!(
-                "each percentage must be in (0.0, 1.0], got {}",
-                b.percentage
-            ));
-        }
-        if !seen.insert(&b.beneficiary) {
-            return Err(format!(
-                "duplicate beneficiary not allowed: {}",
-                b.beneficiary
-            ));
-        }
-    }
-    let sum: DamlDecimal = beneficiaries.iter().map(|b| b.percentage).sum();
-    if sum != one {
-        // Say how to fix it. The comparison is exact Decimal, so an even 3-way
-        // split does not exist and nothing is implicitly left to the decparty —
-        // both are things a proposer discovers at execute otherwise.
-        return Err(format!(
-            "reward beneficiary percentages must sum to exactly 1.0, got {sum}. \
-             The sum is compared as exact Decimal, so balance the last entry by \
-             hand rather than repeating a rounded share. To leave a remainder to \
-             the decparty, list the decparty itself as a beneficiary — nothing is \
-             implicit"
-        ));
-    }
-    Ok(())
 }
 
 /// Request to propose a governance domain action (creates proposal contract)
@@ -1987,33 +1802,8 @@ mod tests {
         assert!(revoke_empty.validate(&cid("gov")).is_err());
     }
 
-    #[test]
-    fn validate_reward_beneficiaries_edge_cases() {
-        // Empty is rejected.
-        assert!(validate_reward_beneficiaries(&[]).is_err());
-
-        // Per-percentage bound is (0.0, 1.0]: 0.0, negative, and > 1.0 all reject.
-        assert!(validate_reward_beneficiaries(&[rb("a", "0.0"), rb("b", "1.0")]).is_err());
-        assert!(validate_reward_beneficiaries(&[rb("a", "-0.5"), rb("b", "1.5")]).is_err());
-        assert!(validate_reward_beneficiaries(&[rb("a", "1.5")]).is_err());
-
-        // A single 1.0 (upper bound inclusive) is accepted.
-        assert!(validate_reward_beneficiaries(&[rb("a", "1.0")]).is_ok());
-
-        // Duplicate beneficiary is rejected even when percentages are otherwise valid.
-        assert!(validate_reward_beneficiaries(&[rb("dup", "0.5"), rb("dup", "0.5")]).is_err());
-
-        // Count boundary: exactly 20 (each 0.05, summing to 1.0) is accepted; 21 rejects.
-        let twenty: Vec<RewardBeneficiary> =
-            (0..20).map(|i| rb(&format!("b{i}"), "0.05")).collect();
-        assert!(validate_reward_beneficiaries(&twenty).is_ok());
-        let twenty_one: Vec<RewardBeneficiary> =
-            (0..21).map(|i| rb(&format!("b{i}"), "0.05")).collect();
-        assert!(validate_reward_beneficiaries(&twenty_one).is_err());
-
-        // Valid two-way split.
-        assert!(validate_reward_beneficiaries(&[rb("a", "0.8"), rb("b", "0.2")]).is_ok());
-    }
+    // `validate_reward_beneficiaries_edge_cases` moved to
+    // `decman_lib::framework::validate::tests` with the validator itself.
 
     fn test_party(prefix: &str) -> anyhow::Result<CantonId> {
         CantonId::parse(&format!("{prefix}::1220{}", "ab".repeat(32)))
